+++ src/main/java/coupon/coupon/service/CouponService.java
@@ -3,15 +3,10 @@ package coupon.coupon.service;
 import coupon.coupon.domain.Coupon;
 import coupon.coupon.domain.CouponCategory;
 import coupon.coupon.domain.CouponEntity;
-import coupon.coupon.domain.MemberCoupon;
-import coupon.coupon.dto.MemberCouponResponse;
 import coupon.coupon.repository.CouponRepository;
-import coupon.coupon.repository.MemberCouponRepository;
 import coupon.coupon.service.validator.CouponDiscountAmountValidator;
 import coupon.coupon.service.validator.CouponMinOrderAmountValidator;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +16,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class CouponService {
 
     private final CouponRepository couponRepository;
-    private final MemberCouponRepository memberCouponRepository;
 
     private final CouponDiscountAmountValidator couponDiscountAmountValidator;
     private final CouponMinOrderAmountValidator couponMinOrderAmountValidator;
@@ -45,36 +39,5 @@ public class CouponService {
     public CouponEntity getCoupon(long couponId) {
         return couponRepository.findById(couponId)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 쿠폰이에요."));
-    }
-
-    @Transactional
-    public long issueCoupon(MemberCoupon memberCoupon) {
-        int sameMemberCouponCount = memberCouponRepository.countAllByMemberIdAndCouponId(
-                memberCoupon.getMemberId(), memberCoupon.getCouponId());
-        if (sameMemberCouponCount > 5) {
-            throw new IllegalArgumentException("동일한 쿠폰은 5장까지 발급할 수 있어요.");
-        }
-        MemberCoupon issuedMemberCoupon = memberCouponRepository.save(memberCoupon);
-        return issuedMemberCoupon.getId();
-    }
-
-    @Transactional(readOnly = true)
-    public List<MemberCouponResponse> getCouponByMember(long memberId) {
-        List<MemberCouponResponse> memberCouponResponses = new ArrayList<>();
-        for (MemberCoupon memberCoupon : memberCouponRepository.findAllByMemberId(memberId)) {
-            CouponEntity coupon = couponRepository.findById(memberCoupon.getCouponId())
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 쿠폰입니다."));
-            MemberCouponResponse memberCouponResponse = new MemberCouponResponse(
-                    coupon.getCouponName(),
-                    coupon.getCouponDiscountAmount(),
-                    coupon.getCouponMinOrderAmount(),
-                    coupon.getCouponCategory().name(),
-                    coupon.getStartAt(),
-                    coupon.getEndAt(),
-                    memberCoupon.isUsed()
-            );
-            memberCouponResponses.add(memberCouponResponse);
-        }
-        return memberCouponResponses;
     }
 }
